+++ changelog.md
@@ -7,6 +7,7 @@ Last change: 08-MAY-2026 CEST
 - Localize: save and load identifications
 - Render: plot profile with adjustable bin width
 - Render GUI: added attribute 'pixelsize' in View for cleaner code
+- Render GUI: use precomputed index blocks for faster circular picked locs calculation
 
 ## 0.10.0